[Decomp][Pet] Cohesion split: extract persistence + spells/talents to Pet*.cpp
