+++ src/game/Object/Pet.cpp
@@ -98,641 +98,6 @@ void Pet::RemoveFromWorld()
     Unit::RemoveFromWorld();
 }
 
-/**
- * @brief Loads a pet from the database for an owner.
- *
- * @param owner The player owning the pet.
- * @param petentry Optional creature entry filter.
- * @param petnumber Optional pet number filter.
- * @param current true to load the current pet.
- * @return true if the pet was loaded successfully; otherwise, false.
- */
-bool Pet::LoadPetFromDB(Player* owner, uint32 petentry, uint32 petnumber, bool current, int32 slot)
-{
-    m_loading = true;
-
-    uint32 ownerid = owner->GetGUIDLow();
-
-    QueryResult* result;
-
-    if (petnumber)
-        // known petnumber entry                   0     1        2(?)     3          4        5      6             7       8       9          10           11         12        13          14                   15                   16                17
-        result = CharacterDatabase.PQuery("SELECT `id`, `entry`, `owner`, `modelid`, `level`, `exp`, `Reactstate`, `slot`, `name`, `renamed`, `curhealth`, `curmana`, `abdata`, `savetime`, `resettalents_cost`, `resettalents_time`, `CreatedBySpell`, `PetType` "
-                                          "FROM `character_pet` WHERE `owner` = '%u' AND `id` = '%u'",
-                                          ownerid, petnumber);
-    else if (current)
-        // current pet (slot 0)                    0     1        2(?)     3          4        5      6             7       8       9          10           11         12        13          14                   15                   16                17
-        result = CharacterDatabase.PQuery("SELECT `id`, `entry`, `owner`, `modelid`, `level`, `exp`, `Reactstate`, `slot`, `name`, `renamed`, `curhealth`, `curmana`, `abdata`, `savetime`, `resettalents_cost`, `resettalents_time`, `CreatedBySpell`, `PetType` "
-                                          "FROM `character_pet` WHERE `owner` = '%u' AND `slot` = '%u'",
-                                          ownerid, PET_SAVE_AS_CURRENT);
-    else if (slot >= 0)
-        // Cata Call Pet 1..N: load the pet at this specific active slot
-        // (0..PET_SLOT_LAST_ACTIVE_SLOT). Caller is Spell::CheckCast for
-        // the hunter Call Pet spell family in a future commit on this
-        // branch; until then this branch is unreachable and the existing
-        // dispatch falls through as before.
-        //                                         0     1        2(?)     3          4        5      6             7       8       9          10           11         12        13          14                   15                   16                17
-        result = CharacterDatabase.PQuery("SELECT `id`, `entry`, `owner`, `modelid`, `level`, `exp`, `Reactstate`, `slot`, `name`, `renamed`, `curhealth`, `curmana`, `abdata`, `savetime`, `resettalents_cost`, `resettalents_time`, `CreatedBySpell`, `PetType` "
-                                          "FROM `character_pet` WHERE `owner` = '%u' AND `slot` = '%u'",
-                                          ownerid, uint32(slot));
-    else if (petentry)
-        // known petentry entry (unique for summoned pet, but non unique for hunter pet (only from current or not stabled pets)
-        //                                         0     1        2(?)     3          4        5      6             7       8       9          10           11         12        13          14                   15                   16                17
-        result = CharacterDatabase.PQuery("SELECT `id`, `entry`, `owner`, `modelid`, `level`, `exp`, `Reactstate`, `slot`, `name`, `renamed`, `curhealth`, `curmana`, `abdata`, `savetime`, `resettalents_cost`, `resettalents_time`, `CreatedBySpell`, `PetType` "
-                                          "FROM `character_pet` WHERE `owner` = '%u' AND `entry` = '%u' AND (`slot` = '%u' OR `slot` > '%u') ",
-                                          ownerid, petentry, PET_SAVE_AS_CURRENT, PET_SAVE_LAST_STABLE_SLOT);
-    else
-        // any current or other non-stabled pet (for hunter "call pet")
-        //                                         0     1        2(?)     3          4        5      6             7       8       9          10           11         12        13          14                   15                   16                17
-        result = CharacterDatabase.PQuery("SELECT `id`, `entry`, `owner`, `modelid`, `level`, `exp`, `Reactstate`, `slot`, `name`, `renamed`, `curhealth`, `curmana`, `abdata`, `savetime`, `resettalents_cost`, `resettalents_time`, `CreatedBySpell`, `PetType` "
-                                          "FROM `character_pet` WHERE `owner` = '%u' AND (`slot` = '%u' OR `slot` > '%u') ",
-                                          ownerid, PET_SAVE_AS_CURRENT, PET_SAVE_LAST_STABLE_SLOT);
-
-    if (!result)
-    {
-        return false;
-    }
-
-    Field* fields = result->Fetch();
-
-    // Record the slot column on the Pet instance. Future commits on this
-    // branch (audit IDs C2 / D-row hunter re-saves) read m_petSlot to keep
-    // each tamed pet parked at its Call Pet 1..N slot across re-saves and
-    // dismisses. Capturing it here (after `result` is known non-null and
-    // the row exists, but before any early-out) means every successful
-    // load -- by petnumber, current-flag, explicit slot, petentry, or
-    // legacy fallback -- gets the same accurate value. Harmless for the
-    // early-out failure paths below: a Pet object that returns false
-    // gets discarded by the caller, m_petSlot included.
-    m_petSlot = int32(fields[7].GetUInt32());
-
-    // update for case of current pet "slot = 0"
-    petentry = fields[1].GetUInt32();
-    if (!petentry)
-    {
-        delete result;
-        return false;
-    }
-
-    CreatureInfo const* creatureInfo = ObjectMgr::GetCreatureTemplate(petentry);
-    if (!creatureInfo)
-    {
-        sLog.outError("Pet entry %u does not exist but used at pet load (owner: %s).", petentry, owner->GetGuidStr().c_str());
-        delete result;
-        return false;
-    }
-
-    uint32 summon_spell_id = fields[16].GetUInt32();
-    SpellEntry const* spellInfo = sSpellStore.LookupEntry(summon_spell_id);
-
-    bool is_temporary_summoned = spellInfo && GetSpellDuration(spellInfo) > 0;
-
-    // check temporary summoned pets like mage water elemental
-    if (current && is_temporary_summoned)
-    {
-        delete result;
-        return false;
-    }
-
-    PetType pet_type = PetType(fields[17].GetUInt8());
-    if (pet_type == HUNTER_PET)
-    {
-        if (!creatureInfo->isTameable(owner->CanTameExoticPets()))
-        {
-            delete result;
-            return false;
-        }
-    }
-
-    uint32 pet_number = fields[0].GetUInt32();
-
-    Map* map = owner->GetMap();
-
-    CreatureCreatePos pos(owner, owner->GetOrientation(), PET_FOLLOW_DIST, PET_FOLLOW_ANGLE);
-
-    uint32 guid = pos.GetMap()->GenerateLocalLowGuid(HIGHGUID_PET);
-    if (!Create(guid, pos, creatureInfo, pet_number))
-    {
-        delete result;
-        return false;
-    }
-
-    setPetType(pet_type);
-    setFaction(owner->getFaction());
-    SetUInt32Value(UNIT_CREATED_BY_SPELL, summon_spell_id);
-
-    // reget for sure use real creature info selected for Pet at load/creating
-    CreatureInfo const* cinfo = GetCreatureInfo();
-    if (cinfo->CreatureType == CREATURE_TYPE_CRITTER)
-    {
-        AIM_Initialize();
-        pos.GetMap()->Add((Creature*)this);
-        delete result;
-        return true;
-    }
-
-    m_charmInfo->SetPetNumber(pet_number, isControlled());
-
-    SetOwnerGuid(owner->GetObjectGuid());
-    SetDisplayId(fields[3].GetUInt32());
-    SetNativeDisplayId(fields[3].GetUInt32());
-    uint32 petlevel = fields[4].GetUInt32();
-    SetUInt32Value(UNIT_NPC_FLAGS, UNIT_NPC_FLAG_NONE);
-    SetName(fields[8].GetString());
-
-    // Defensive name fallback. The starter pet rows created by
-    // WorldSession::HandleCharCreateOpcode (CharacterHandler.cpp) write
-    // `name = ' '` (single space) into character_pet for every race's
-    // hunter starter and the warlock imp. An empty / whitespace-only
-    // name renders client-side as "Unknown's Pet" in unit frames and
-    // produces blank-name rows in the stable list. Any future code path
-    // that persists a Pet row without setting a name lands here too --
-    // this fallback mirrors the family-name logic in CreateBaseAtCreature
-    // (PR #135) so the load path is consistent with the tame path.
-    {
-        bool nameIsBlank = true;
-        for (char const* p = m_name.c_str(); *p; ++p)
-        {
-            if (*p != ' ' && *p != '\t' && *p != '\r' && *p != '\n')
-            {
-                nameIsBlank = false;
-                break;
-            }
-        }
-        if (nameIsBlank)
-        {
-            std::string fallback;
-            if (CreatureFamilyEntry const* cFamily = sCreatureFamilyStore.LookupEntry(creatureInfo->Family))
-            {
-                fallback = cFamily->Name[sWorld.GetDefaultDbcLocale()];
-            }
-            if (fallback.empty() && creatureInfo->Name)
-            {
-                fallback = creatureInfo->Name;
-            }
-            if (!fallback.empty())
-            {
-                SetName(fallback);
-            }
-        }
-    }
-
-    SetByteValue(UNIT_FIELD_BYTES_2, 1, UNIT_BYTE2_FLAG_SUPPORTABLE | UNIT_BYTE2_FLAG_AURAS);
-    SetUInt32Value(UNIT_FIELD_FLAGS, UNIT_FLAG_PVP_ATTACKABLE);
-
-    if (getPetType() == HUNTER_PET)
-    {
-        SetByteFlag(UNIT_FIELD_BYTES_2, 2, fields[9].GetBool() ? UNIT_CAN_BE_ABANDONED : UNIT_CAN_BE_RENAMED | UNIT_CAN_BE_ABANDONED);
-        SetPowerType(POWER_FOCUS);
-    }
-    else if (getPetType() != SUMMON_PET)
-        sLog.outError("Pet have incorrect type (%u) for pet loading.", getPetType());
-
-    if (owner->IsPvP())
-    {
-        SetPvP(true);
-    }
-
-    if (owner->IsFFAPvP())
-    {
-        SetFFAPvP(true);
-    }
-
-    SetCanModifyStats(true);
-    InitStatsForLevel(petlevel);
-    InitTalentForLevel();                                   // set original talents points before spell loading
-
-    SetUInt32Value(UNIT_FIELD_PET_NAME_TIMESTAMP, uint32(time(NULL)));
-    SetUInt32Value(UNIT_FIELD_PETEXPERIENCE, fields[5].GetUInt32());
-    SetCreatorGuid(owner->GetObjectGuid());
-
-    m_charmInfo->SetReactState(ReactStates(fields[6].GetUInt8()));
-
-    uint32 savedhealth = fields[10].GetUInt32();
-    uint32 savedpower = fields[11].GetUInt32();
-
-    // set current pet as current
-    // 0=current
-    // 1..MAX_PET_STABLES in stable slot
-    // PET_SAVE_NOT_IN_SLOT(100) = not stable slot (summoning))
-    //
-    // The legacy auto-promote block below moves whichever pet we just
-    // loaded into PET_SAVE_AS_CURRENT (slot 0) and bumps any previous
-    // slot-0 pet to PET_SAVE_NOT_IN_SLOT (=100). That was correct under
-    // the WotLK 1-active-pet model where loading a stabled pet implied
-    // making it the active one and demoting the old current.
-    //
-    // For Cata HUNTER_PET the same code path is destructive: every pet
-    // permanently lives in its assigned Call Pet 1..N slot 0..4 and
-    // loading one must NOT shift it. Without this gate,
-    // ResummonTemporaryUnsummonedIfAny (mount, transport, vehicle exit)
-    // would force the resumed pet back to slot 0 and silently bump the
-    // pet that owns slot 0 out into orphan land, breaking the next
-    // Call Pet 1 cast. This was the bug audit row C2 in
-    // MANGOS/PET_SAVE_CALLSITE_AUDIT.md and the gap closed PR #137 left
-    // open when it was reverted.
-    //
-    // SUMMON_PET / GUARDIAN_PET / MINI_PET keep the legacy auto-promote
-    // -- they have no Call Pet 1..N concept and the warlock pet pool
-    // still expects loading from slot 100 to mean "becomes active."
-    if (getPetType() != HUNTER_PET && fields[7].GetUInt32() != 0)
-    {
-        CharacterDatabase.BeginTransaction();
-
-        static SqlStatementID id_1;
-        static SqlStatementID id_2;
-
-        SqlStatement stmt = CharacterDatabase.CreateStatement(id_1, "UPDATE `character_pet` SET `slot` = ? WHERE `owner` = ? AND `slot` = ? AND `id` <> ?");
-        stmt.PExecute(uint32(PET_SAVE_NOT_IN_SLOT), ownerid, uint32(PET_SAVE_AS_CURRENT), m_charmInfo->GetPetNumber());
-
-        stmt = CharacterDatabase.CreateStatement(id_2, "UPDATE `character_pet` SET `slot` = ? WHERE `owner` = ? AND `id` = ?");
-        stmt.PExecute(uint32(PET_SAVE_AS_CURRENT), ownerid, m_charmInfo->GetPetNumber());
-
-        CharacterDatabase.CommitTransaction();
-    }
-
-    // load action bar, if data broken will fill later by default spells.
-    if (!is_temporary_summoned)
-    {
-        m_charmInfo->LoadPetActionBar(fields[12].GetCppString());
-    }
-
-    // since last save (in seconds)
-    uint32 timediff = uint32(time(NULL) - fields[13].GetUInt64());
-
-    m_resetTalentsCost = fields[14].GetUInt32();
-    m_resetTalentsTime = fields[15].GetUInt64();
-
-    delete result;
-
-    // load spells/cooldowns/auras
-    _LoadAuras(timediff);
-
-    // init AB
-    if (is_temporary_summoned)
-    {
-        // Temporary summoned pets always have initial spell list at load
-        InitPetCreateSpells();
-    }
-    else
-    {
-        LearnPetPassives();
-        CastPetAuras(current);
-        CastOwnerTalentAuras();
-    }
-
-    Powers powerType = GetPowerType();;
-
-    SetHealth(savedhealth > GetMaxHealth() ? GetMaxHealth() : savedhealth);
-    SetPower(powerType, savedpower > GetMaxPower(powerType) ? GetMaxPower(powerType) : savedpower);
-
-    if (getPetType() == HUNTER_PET && savedhealth <= 0)
-    {
-        SetDeathState(JUST_DIED);
-    }
-
-    map->Add((Creature*)this);
-    AIM_Initialize();
-
-    // Spells should be loaded after pet is added to map, because in CheckCast is check on it
-    _LoadSpells();
-    InitLevelupSpellsForLevel();
-
-    CleanupActionBar();                                     // remove unknown spells from action bar after load
-
-    _LoadSpellCooldowns();
-
-    owner->SetPet(this);                                    // in DB stored only full controlled creature
-    DEBUG_LOG("New Pet has guid %u", GetGUIDLow());
-
-    if (owner->GetTypeId() == TYPEID_PLAYER)
-    {
-        ((Player*)owner)->PetSpellInitialize();
-        if (((Player*)owner)->GetGroup())
-        {
-            ((Player*)owner)->SetGroupUpdateFlag(GROUP_UPDATE_PET);
-        }
-
-        ((Player*)owner)->SendTalentsInfoData(true);
-    }
-
-    if (owner->GetTypeId() == TYPEID_PLAYER && getPetType() == HUNTER_PET)
-    {
-        result = CharacterDatabase.PQuery("SELECT `genitive`, `dative`, `accusative`, `instrumental`, `prepositional` FROM `character_pet_declinedname` WHERE `owner` = '%u' AND `id` = '%u'", owner->GetGUIDLow(), GetCharmInfo()->GetPetNumber());
-
-        if (result)
-        {
-            delete m_declinedname;
-            m_declinedname = new DeclinedName;
-
-            Field* fields2 = result->Fetch();
-            for (int i = 0; i < MAX_DECLINED_NAME_CASES; ++i)
-            {
-                m_declinedname->name[i] = fields2[i].GetCppString();
-            }
-
-            delete result;
-        }
-    }
-
-    m_loading = false;
-
-    SynchronizeLevelWithOwner();
-    return true;
-}
-
-/**
- * @brief Saves the pet to the database using the specified save mode.
- *
- * @param mode The pet save mode to apply.
- */
-void Pet::SavePetToDB(PetSaveMode mode)
-{
-    if (!GetEntry())
-    {
-        return;
-    }
-
-    // save only fully controlled creature
-    if (!isControlled())
-    {
-        return;
-    }
-
-    // not save not player pets
-    if (!GetOwnerGuid().IsPlayer())
-    {
-        return;
-    }
-
-    Player* pOwner = (Player*)GetOwner();
-    if (!pOwner)
-    {
-        return;
-    }
-
-    // Cata multi-pet allocator. Resolves PET_SAVE_NEW_PET (== 102) into
-    // a concrete active-roster slot 0..PET_SLOT_LAST_ACTIVE_SLOT by
-    // scanning character_pet for the first vacant slot in that range.
-    // Updates `mode` and `m_petSlot` in place so the rest of the
-    // function continues unchanged. If the roster is full, leaves
-    // m_petSlot at PET_SAVE_NOT_IN_SLOT and bails out without writing
-    // any row -- the caller (EffectTameCreature, future commit) checks
-    // GetSlot() after the call to detect failure and roll the tame back.
-    //
-    // No caller passes PET_SAVE_NEW_PET in this commit; the block
-    // becomes reachable in step 10 of the audit. Adding the allocator
-    // ahead of its caller keeps each diff small enough to review on
-    // its own.
-    if (mode == PET_SAVE_NEW_PET)
-    {
-        uint32 ownerLowForNewPet = GetOwnerGuid().GetCounter();
-        bool occupied[PET_SLOT_LAST_ACTIVE_SLOT + 1] = {};
-        if (QueryResult* used = CharacterDatabase.PQuery(
-                "SELECT `slot` FROM `character_pet` WHERE `owner` = '%u' AND `slot` <= '%u'",
-                ownerLowForNewPet, uint32(PET_SLOT_LAST_ACTIVE_SLOT)))
-        {
-            do
-            {
-                uint32 s = used->Fetch()[0].GetUInt32();
-                if (s <= uint32(PET_SLOT_LAST_ACTIVE_SLOT))
-                {
-                    occupied[s] = true;
-                }
-            }
-            while (used->NextRow());
-            delete used;
-        }
-        int32 freeSlot = -1;
-        for (int32 s = 0; s <= PET_SLOT_LAST_ACTIVE_SLOT; ++s)
-        {
-            if (!occupied[s])
-            {
-                freeSlot = s;
-                break;
-            }
-        }
-        if (freeSlot < 0)
-        {
-            // Roster full -- do not persist anything. m_petSlot value
-            // signals "no slot allocated" to the caller.
-            m_petSlot = int32(PET_SAVE_NOT_IN_SLOT);
-            return;
-        }
-        m_petSlot = freeSlot;
-        mode = static_cast<PetSaveMode>(freeSlot);
-    }
-    else if (getPetType() == HUNTER_PET
-             && (mode == PET_SAVE_AS_CURRENT || mode == PET_SAVE_NOT_IN_SLOT)
-             && m_petSlot >= 0
-             && m_petSlot <= int32(PET_SLOT_LAST_ACTIVE_SLOT))
-    {
-        // Cata multi-pet re-save path. Every hunter pet stays parked in
-        // its Call Pet 1..N slot for the entire character lifetime,
-        // including across dismiss, logout, level-up, feed, range
-        // unsummon and visibility teardown. The pre-Cata code routed
-        // PET_SAVE_AS_CURRENT (=0) and PET_SAVE_NOT_IN_SLOT (=100)
-        // directly through to the INSERT, evicting the pet from its
-        // assigned slot and breaking the next Call Pet N cast with
-        // NO_PET.
-        //
-        // Rewrite mode in place so the rest of SavePetToDB writes
-        // back to the pet's known slot. Legacy WotLK paths that have
-        // never set m_petSlot (m_petSlot == -1, e.g. a fresh Pet*
-        // about to be saved for the first time -- though tame goes
-        // through PET_SAVE_NEW_PET in step 10) still fall through to
-        // the original mode-as-passed behaviour. SUMMON_PET /
-        // GUARDIAN_PET also fall through (gated on HUNTER_PET).
-        //
-        // Audit row IDs: D1 (EffectDismissPet), D2 (mount/transport
-        // temp-unsummon), D4 (out-of-range), D5 (talent reset),
-        // D6 (visibility teardown), D7 (GM .pet unsummon),
-        // S4 (Player::SaveToDB periodic).
-        //
-        // Until step 10 ships, no hunter has a pet with m_petSlot
-        // in 0..PET_SLOT_LAST_ACTIVE_SLOT (taming still goes through
-        // PET_SAVE_AS_CURRENT and m_petSlot stays at -1 because there
-        // is no load path to populate it). So this branch is dormant
-        // in practice today and engages once step 10's tame change
-        // starts populating m_petSlot.
-        mode = static_cast<PetSaveMode>(m_petSlot);
-    }
-
-    // current/stable/not_in_slot
-    if (mode >= PET_SAVE_AS_CURRENT)
-    {
-        // reagents must be returned before save call
-        if (mode == PET_SAVE_REAGENTS)
-        {
-            // Hunter Pets always save as current if dismissed or unsummoned due to range/etc.
-            if (getPetType() == HUNTER_PET)
-            {
-                mode = PET_SAVE_AS_CURRENT;
-            }
-            else
-            {
-                mode = PET_SAVE_NOT_IN_SLOT;
-            }
-        }
-        // not save pet as current if another pet temporary unsummoned
-        else if (mode == PET_SAVE_AS_CURRENT && pOwner->GetTemporaryUnsummonedPetNumber() &&
-                 pOwner->GetTemporaryUnsummonedPetNumber() != m_charmInfo->GetPetNumber())
-        {
-            // pet will lost anyway at restore temporary unsummoned
-            if (getPetType() == HUNTER_PET)
-            {
-                return;
-            }
-
-            // for warlock case
-            mode = PET_SAVE_NOT_IN_SLOT;
-        }
-
-        uint32 curhealth = GetHealth();
-        uint32 curpower = GetPower(GetPowerType());
-
-        // stable and not in slot saves
-        if (mode != PET_SAVE_AS_CURRENT)
-        {
-            RemoveAllAuras();
-        }
-
-        // save pet's data as one single transaction
-        CharacterDatabase.BeginTransaction();
-        _SaveSpells();
-        _SaveSpellCooldowns();
-        _SaveAuras();
-
-        uint32 ownerLow = GetOwnerGuid().GetCounter();
-        // remove current data
-        static SqlStatementID delPet ;
-        static SqlStatementID insPet ;
-
-        SqlStatement stmt = CharacterDatabase.CreateStatement(delPet, "DELETE FROM `character_pet` WHERE `owner` = ? AND `id` = ?");
-        stmt.PExecute(ownerLow, m_charmInfo->GetPetNumber());
-
-        // prevent duplicate using slot (except PET_SAVE_NOT_IN_SLOT)
-        if (mode <= PET_SAVE_LAST_STABLE_SLOT)
-        {
-            static SqlStatementID updPet ;
-
-            stmt = CharacterDatabase.CreateStatement(updPet, "UPDATE `character_pet` SET `slot` = ? WHERE `owner` = ? AND `slot` = ?");
-            stmt.PExecute(uint32(PET_SAVE_NOT_IN_SLOT), ownerLow, uint32(mode));
-        }
-
-        // Cata multi-pet hunter cleanup: reap orphan rows
-        // (slot > PET_SAVE_LAST_STABLE_SLOT) only -- the
-        // slot 0..PET_SAVE_LAST_STABLE_SLOT range is the Call Pet 1..N
-        // roster and must be preserved across every save.
-        //
-        // The legacy WHERE clause was `slot = PET_SAVE_AS_CURRENT OR
-        // slot > PET_SAVE_LAST_STABLE_SLOT`, which destroyed every
-        // other tamed pet at slot 0 whenever the active pet was
-        // re-saved (dismiss, level-up, logout, taming a new one).
-        // That was the data-loss root surfaced when PR #137 was
-        // tested in-game and reverted -- the fix is here, in the
-        // WHERE clause, not in the per-mode intercepts.
-        //
-        // `id <> ?` excludes the pet we are about to INSERT below so
-        // that even if its previous row was bumped to NOT_IN_SLOT by
-        // the UPDATE above, we do not immediately delete the new row.
-        //
-        // No behaviour change for SUMMON_PET / GUARDIAN_PET (the
-        // outer `getPetType() == HUNTER_PET` guard already excluded
-        // them). No behaviour change for any current hunter that has
-        // a single pet at slot 0 -- there are no orphan rows to reap,
-        // so the DELETE is a no-op on a healthy single-pet DB. The
-        // cleanup re-engages only once multi-pet hunters exist
-        // (after step 10 of the audit ships).
-        if (getPetType() == HUNTER_PET && (mode == PET_SAVE_AS_CURRENT || mode > PET_SAVE_LAST_STABLE_SLOT))
-        {
-            static SqlStatementID del ;
-
-            stmt = CharacterDatabase.CreateStatement(del, "DELETE FROM `character_pet` WHERE `owner` = ? AND `slot` > ? AND `id` <> ?");
-            stmt.PExecute(ownerLow, uint32(PET_SAVE_LAST_STABLE_SLOT), m_charmInfo->GetPetNumber());
-        }
-
-        // save pet
-        SqlStatement savePet = CharacterDatabase.CreateStatement(insPet, "INSERT INTO `character_pet` "
-        "(`id`, `entry`,  `owner`, `modelid`, `level`, `exp`, `Reactstate`, `slot`, `name`, `renamed`, `curhealth`, `curmana`, `abdata`, `savetime`, `resettalents_cost`, `resettalents_time`, `CreatedBySpell`, `PetType`) "
-             "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
-
-        savePet.addUInt32(m_charmInfo->GetPetNumber());
-        savePet.addUInt32(GetEntry());
-        savePet.addUInt32(ownerLow);
-        savePet.addUInt32(GetNativeDisplayId());
-        savePet.addUInt32(getLevel());
-        savePet.addUInt32(GetUInt32Value(UNIT_FIELD_PETEXPERIENCE));
-        savePet.addUInt32(uint32(m_charmInfo->GetReactState()));
-        savePet.addUInt32(uint32(mode));
-        savePet.addString(m_name.c_str());
-        savePet.addUInt32(uint32(HasByteFlag(UNIT_FIELD_BYTES_2, 2, UNIT_CAN_BE_RENAMED) ? 0 : 1));
-        savePet.addUInt32(curhealth);
-        savePet.addUInt32(curpower);
-
-        std::ostringstream ss;
-        for (uint32 i = ACTION_BAR_INDEX_START; i < ACTION_BAR_INDEX_END; ++i)
-        {
-            ss << uint32(m_charmInfo->GetActionBarEntry(i)->GetType()) << " "
-               << uint32(m_charmInfo->GetActionBarEntry(i)->GetAction()) << " ";
-        };
-        savePet.addString(ss);
-
-        savePet.addUInt64(uint64(time(NULL)));
-        savePet.addUInt32(uint32(m_resetTalentsCost));
-        savePet.addUInt64(uint64(m_resetTalentsTime));
-        savePet.addUInt32(GetUInt32Value(UNIT_CREATED_BY_SPELL));
-        savePet.addUInt32(uint32(getPetType()));
-
-        savePet.Execute();
-        CharacterDatabase.CommitTransaction();
-    }
-    else
-    {
-        RemoveAllAuras(AURA_REMOVE_BY_DELETE);
-        DeleteFromDB(m_charmInfo->GetPetNumber());
-    }
-}
-
-/**
- * @brief Deletes pet-related database records.
- *
- * @param guidlow The pet number identifier.
- * @param separate_transaction true to wrap deletion in its own transaction.
- */
-void Pet::DeleteFromDB(uint32 guidlow, bool separate_transaction)
-{
-    if (separate_transaction)
-    {
-        CharacterDatabase.BeginTransaction();
-    }
-
-    static SqlStatementID delPet ;
-    static SqlStatementID delDeclName ;
-    static SqlStatementID delAuras ;
-    static SqlStatementID delSpells ;
-    static SqlStatementID delSpellCD ;
-
-    SqlStatement stmt = CharacterDatabase.CreateStatement(delPet, "DELETE FROM `character_pet` WHERE `id` = ?");
-    stmt.PExecute(guidlow);
-
-    stmt = CharacterDatabase.CreateStatement(delDeclName, "DELETE FROM `character_pet_declinedname` WHERE `id` = ?");
-    stmt.PExecute(guidlow);
-
-    stmt = CharacterDatabase.CreateStatement(delAuras, "DELETE FROM `pet_aura` WHERE `guid` = ?");
-    stmt.PExecute(guidlow);
-
-    stmt = CharacterDatabase.CreateStatement(delSpells, "DELETE FROM `pet_spell` WHERE `guid` = ?");
-    stmt.PExecute(guidlow);
-
-    stmt = CharacterDatabase.CreateStatement(delSpellCD, "DELETE FROM `pet_spell_cooldown` WHERE `guid` = ?");
-    stmt.PExecute(guidlow);
-
-    if (separate_transaction)
-    {
-        CharacterDatabase.CommitTransaction();
-    }
-}
 
 /**
  * @brief Updates the pet death state and related pet-specific behavior.
@@ -1562,1091 +927,6 @@ uint32 Pet::GetCurrentFoodBenefitLevel(uint32 itemlevel)
     }
 }
 
-/**
- * @brief Loads saved pet spell cooldowns from the database.
- */
-void Pet::_LoadSpellCooldowns()
-{
-    m_CreatureSpellCooldowns.clear();
-    m_CreatureCategoryCooldowns.clear();
-
-    QueryResult* result = CharacterDatabase.PQuery("SELECT `spell`,`time` FROM `pet_spell_cooldown` WHERE `guid` = '%u'", m_charmInfo->GetPetNumber());
-
-    if (result)
-    {
-        time_t curTime = time(NULL);
-
-        WorldPacket data(SMSG_SPELL_COOLDOWN, (8 + 1 + size_t(result->GetRowCount()) * 8));
-        data << ObjectGuid(GetObjectGuid());
-        data << uint8(0x0);                                 // flags (0x1, 0x2)
-
-        do
-        {
-            Field* fields = result->Fetch();
-
-            uint32 spell_id = fields[0].GetUInt32();
-            time_t db_time  = (time_t)fields[1].GetUInt64();
-
-            if (!sSpellStore.LookupEntry(spell_id))
-            {
-                sLog.outError("Pet %u have unknown spell %u in `pet_spell_cooldown`, skipping.", m_charmInfo->GetPetNumber(), spell_id);
-                continue;
-            }
-
-            // skip outdated cooldown
-            if (db_time <= curTime)
-            {
-                continue;
-            }
-
-            data << uint32(spell_id);
-            data << uint32(uint32(db_time - curTime)*IN_MILLISECONDS);
-
-            _AddCreatureSpellCooldown(spell_id, db_time);
-
-            DEBUG_LOG("Pet (Number: %u) spell %u cooldown loaded (%u secs).", m_charmInfo->GetPetNumber(), spell_id, uint32(db_time - curTime));
-        }
-        while (result->NextRow());
-
-        delete result;
-
-        if (!m_CreatureSpellCooldowns.empty() && GetOwner())
-        {
-            ((Player*)GetOwner())->GetSession()->SendPacket(&data);
-        }
-    }
-}
-
-/**
- * @brief Saves active pet spell cooldowns to the database.
- */
-void Pet::_SaveSpellCooldowns()
-{
-    static SqlStatementID delSpellCD ;
-    static SqlStatementID insSpellCD ;
-
-    SqlStatement stmt = CharacterDatabase.CreateStatement(delSpellCD, "DELETE FROM `pet_spell_cooldown` WHERE `guid` = ?");
-    stmt.PExecute(m_charmInfo->GetPetNumber());
-
-    time_t curTime = time(NULL);
-
-    // remove oudated and save active
-    for (CreatureSpellCooldowns::iterator itr = m_CreatureSpellCooldowns.begin(); itr != m_CreatureSpellCooldowns.end();)
-    {
-        if (itr->second <= curTime)
-        {
-            m_CreatureSpellCooldowns.erase(itr++);
-        }
-        else
-        {
-            stmt = CharacterDatabase.CreateStatement(insSpellCD, "INSERT INTO `pet_spell_cooldown` (`guid`,`spell`,`time`) VALUES (?, ?, ?)");
-            stmt.PExecute(m_charmInfo->GetPetNumber(), itr->first, uint64(itr->second));
-            ++itr;
-        }
-    }
-}
-
-/**
- * @brief Loads pet spells from the database.
- */
-void Pet::_LoadSpells()
-{
-    QueryResult* result = CharacterDatabase.PQuery("SELECT `spell`,`active` FROM `pet_spell` WHERE `guid` = '%u'", m_charmInfo->GetPetNumber());
-
-    if (result)
-    {
-        do
-        {
-            Field* fields = result->Fetch();
-
-            addSpell(fields[0].GetUInt32(), ActiveStates(fields[1].GetUInt8()), PETSPELL_UNCHANGED);
-        }
-        while (result->NextRow());
-
-        delete result;
-    }
-}
-
-/**
- * @brief Saves pet spells to the database.
- */
-void Pet::_SaveSpells()
-{
-    static SqlStatementID delSpell ;
-    static SqlStatementID insSpell ;
-
-    for (PetSpellMap::iterator itr = m_spells.begin(), next = m_spells.begin(); itr != m_spells.end(); itr = next)
-    {
-        ++next;
-
-        // prevent saving family passives to DB
-        if (itr->second.type == PETSPELL_FAMILY)
-        {
-            continue;
-        }
-
-        switch (itr->second.state)
-        {
-            case PETSPELL_REMOVED:
-            {
-                SqlStatement stmt = CharacterDatabase.CreateStatement(delSpell, "DELETE FROM `pet_spell` WHERE `guid` = ? AND `spell` = ?");
-                stmt.PExecute(m_charmInfo->GetPetNumber(), itr->first);
-                m_spells.erase(itr);
-            }
-            continue;
-            case PETSPELL_CHANGED:
-            {
-                SqlStatement stmt = CharacterDatabase.CreateStatement(delSpell, "DELETE FROM `pet_spell` WHERE `guid` = ? AND `spell` = ?");
-                stmt.PExecute(m_charmInfo->GetPetNumber(), itr->first);
-
-                stmt = CharacterDatabase.CreateStatement(insSpell, "INSERT INTO `pet_spell` (`guid`,`spell`,`active`) VALUES (?, ?, ?)");
-                stmt.PExecute(m_charmInfo->GetPetNumber(), itr->first, uint32(itr->second.active));
-            }
-            break;
-            case PETSPELL_NEW:
-            {
-                SqlStatement stmt = CharacterDatabase.CreateStatement(insSpell, "INSERT INTO `pet_spell` (`guid`,`spell`,`active`) VALUES (?, ?, ?)");
-                stmt.PExecute(m_charmInfo->GetPetNumber(), itr->first, uint32(itr->second.active));
-            }
-            break;
-            case PETSPELL_UNCHANGED:
-                continue;
-        }
-
-        itr->second.state = PETSPELL_UNCHANGED;
-    }
-}
-
-/**
- * @brief Loads persistent pet auras from the database.
- *
- * @param timediff Time elapsed since last save, in seconds.
- */
-void Pet::_LoadAuras(uint32 timediff)
-{
-    RemoveAllAuras();
-
-    QueryResult* result = CharacterDatabase.PQuery("SELECT `caster_guid`,`item_guid`,`spell`,`stackcount`,`remaincharges`,`basepoints0`,`basepoints1`,`basepoints2`,`periodictime0`,`periodictime1`,`periodictime2`,`maxduration`,`remaintime`,`effIndexMask` FROM `pet_aura` WHERE `guid` = '%u'", m_charmInfo->GetPetNumber());
-
-    if (result)
-    {
-        do
-        {
-            Field* fields = result->Fetch();
-            ObjectGuid casterGuid = ObjectGuid(fields[0].GetUInt64());
-            uint32 item_lowguid = fields[1].GetUInt32();
-            uint32 spellid = fields[2].GetUInt32();
-            uint32 stackcount = fields[3].GetUInt32();
-            uint32 remaincharges = fields[4].GetUInt32();
-            int32  damage[MAX_EFFECT_INDEX];
-            uint32 periodicTime[MAX_EFFECT_INDEX];
-
-            for (int32 i = 0; i < MAX_EFFECT_INDEX; ++i)
-            {
-                damage[i] = fields[i + 5].GetInt32();
-                periodicTime[i] = fields[i + 8].GetUInt32();
-            }
-
-            int32 maxduration = fields[11].GetInt32();
-            int32 remaintime = fields[12].GetInt32();
-            uint32 effIndexMask = fields[13].GetUInt32();
-
-            SpellEntry const* spellproto = sSpellStore.LookupEntry(spellid);
-            if (!spellproto)
-            {
-                sLog.outError("Unknown spell (spellid %u), ignore.", spellid);
-                continue;
-            }
-
-            // do not load single target auras (unless they were cast by the player)
-            if (casterGuid != GetObjectGuid() && IsSingleTargetSpell(spellproto))
-            {
-                continue;
-            }
-
-            if (remaintime != -1 && !IsPositiveSpell(spellproto))
-            {
-                if (remaintime / IN_MILLISECONDS <= int32(timediff))
-                {
-                    continue;
-                }
-
-                remaintime -= timediff * IN_MILLISECONDS;
-            }
-
-            // prevent wrong values of remaincharges
-            uint32 procCharges = spellproto->GetProcCharges();
-            if (procCharges)
-            {
-                if (remaincharges <= 0 || remaincharges > procCharges)
-                {
-                    remaincharges = procCharges;
-                }
-            }
-            else
-            {
-                remaincharges = 0;
-            }
-
-            uint32 defstackamount = spellproto->GetStackAmount();
-            if (!defstackamount)
-            {
-                stackcount = 1;
-            }
-            else if (defstackamount < stackcount)
-            {
-                stackcount = defstackamount;
-            }
-            else if (!stackcount)
-            {
-                stackcount = 1;
-            }
-
-            SpellAuraHolder* holder = CreateSpellAuraHolder(spellproto, this, NULL);
-            holder->SetLoadedState(casterGuid, ObjectGuid(HIGHGUID_ITEM, item_lowguid), stackcount, remaincharges, maxduration, remaintime);
-
-            for (int32 i = 0; i < MAX_EFFECT_INDEX; ++i)
-            {
-                if ((effIndexMask & (1 << i)) == 0)
-                {
-                    continue;
-                }
-
-                Aura* aura = CreateAura(spellproto, SpellEffectIndex(i), NULL, holder, this);
-                if (!damage[i])
-                {
-                    damage[i] = aura->GetModifier()->m_amount;
-                }
-
-                aura->SetLoadedState(damage[i], periodicTime[i]);
-                holder->AddAura(aura, SpellEffectIndex(i));
-            }
-
-            if (!holder->IsEmptyHolder())
-            {
-                AddSpellAuraHolder(holder);
-            }
-            else
-            {
-                delete holder;
-            }
-        }
-        while (result->NextRow());
-
-        delete result;
-    }
-}
-
-/**
- * @brief Saves persistent pet auras to the database.
- */
-void Pet::_SaveAuras()
-{
-    static SqlStatementID delAuras ;
-    static SqlStatementID insAuras ;
-
-    SqlStatement stmt = CharacterDatabase.CreateStatement(delAuras, "DELETE FROM `pet_aura` WHERE `guid` = ?");
-    stmt.PExecute(m_charmInfo->GetPetNumber());
-
-    SpellAuraHolderMap const& auraHolders = GetSpellAuraHolderMap();
-
-    if (auraHolders.empty())
-    {
-        return;
-    }
-
-    stmt = CharacterDatabase.CreateStatement(insAuras, "INSERT INTO `pet_aura` (`guid`, `caster_guid`, `item_guid`, `spell`, `stackcount`, `remaincharges`, "
-            "`basepoints0`, `basepoints1`, `basepoints2`, `periodictime0`, `periodictime1`, `periodictime2`, `maxduration`, `remaintime`, `effIndexMask`) "
-            "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
-
-    for (SpellAuraHolderMap::const_iterator itr = auraHolders.begin(); itr != auraHolders.end(); ++itr)
-    {
-        SpellAuraHolder* holder = itr->second;
-
-        bool save = true;
-        for (int32 j = 0; j < MAX_EFFECT_INDEX; ++j)
-        {
-            SpellEntry const* spellInfo = holder->GetSpellProto();
-            SpellEffectEntry const* effectEntry = spellInfo->GetSpellEffect(SpellEffectIndex(j));
-            if (!effectEntry)
-            {
-                continue;
-            }
-
-            if (effectEntry->EffectApplyAuraName == SPELL_AURA_MOD_STEALTH ||
-                effectEntry->Effect == SPELL_EFFECT_APPLY_AREA_AURA_OWNER ||
-                effectEntry->Effect == SPELL_EFFECT_APPLY_AREA_AURA_PET )
-            {
-                save = false;
-                break;
-            }
-        }
-
-        // skip all holders from spells that are passive or channeled
-        // do not save single target holders (unless they were cast by the player)
-        if (save && !holder->IsPassive() && !IsChanneledSpell(holder->GetSpellProto()) && (holder->GetCasterGuid() == GetObjectGuid() || holder->GetTrackedAuraType() != TRACK_AURA_TYPE_NOT_TRACKED))
-        {
-            int32  damage[MAX_EFFECT_INDEX];
-            uint32 periodicTime[MAX_EFFECT_INDEX];
-            uint32 effIndexMask = 0;
-
-            for (uint32 i = 0; i < MAX_EFFECT_INDEX; ++i)
-            {
-                damage[i] = 0;
-                periodicTime[i] = 0;
-
-                if (Aura* aur = holder->GetAuraByEffectIndex(SpellEffectIndex(i)))
-                {
-                    // don't save not own area auras
-                    if (aur->IsAreaAura() && holder->GetCasterGuid() != GetObjectGuid())
-                    {
-                        continue;
-                    }
-
-                    damage[i] = aur->GetModifier()->m_amount;
-                    periodicTime[i] = aur->GetModifier()->periodictime;
-                    effIndexMask |= (1 << i);
-                }
-            }
-
-            if (!effIndexMask)
-            {
-                continue;
-            }
-
-            stmt.addUInt32(m_charmInfo->GetPetNumber());
-            stmt.addUInt64(holder->GetCasterGuid().GetRawValue());
-            stmt.addUInt32(holder->GetCastItemGuid().GetCounter());
-            stmt.addUInt32(holder->GetId());
-            stmt.addUInt32(holder->GetStackAmount());
-            stmt.addUInt8(holder->GetAuraCharges());
-
-            for (uint32 i = 0; i < MAX_EFFECT_INDEX; ++i)
-            {
-                stmt.addInt32(damage[i]);
-            }
-
-            for (uint32 i = 0; i < MAX_EFFECT_INDEX; ++i)
-            {
-                stmt.addUInt32(periodicTime[i]);
-            }
-
-            stmt.addInt32(holder->GetAuraMaxDuration());
-            stmt.addInt32(holder->GetAuraDuration());
-            stmt.addUInt32(effIndexMask);
-            stmt.Execute();
-        }
-    }
-}
-
-/**
- * @brief Adds a spell to the pet spellbook.
- *
- * @param spell_id The spell to add.
- * @param active The desired active state.
- * @param state The persistence state of the spell.
- * @param type The pet spell category.
- * @return true if the spell was added; otherwise, false.
- */
-bool Pet::addSpell(uint32 spell_id, ActiveStates active /*= ACT_DECIDE*/, PetSpellState state /*= PETSPELL_NEW*/, PetSpellType type /*= PETSPELL_NORMAL*/)
-{
-    SpellEntry const* spellInfo = sSpellStore.LookupEntry(spell_id);
-    if (!spellInfo)
-    {
-        // do pet spell book cleanup
-        if (state == PETSPELL_UNCHANGED)                    // spell load case
-        {
-            sLog.outError("Pet::addSpell: nonexistent in SpellStore spell #%u request, deleting for all pets in `pet_spell`.", spell_id);
-            CharacterDatabase.PExecute("DELETE FROM `pet_spell` WHERE `spell` = '%u'", spell_id);
-        }
-        else
-        {
-            sLog.outError("Pet::addSpell: nonexistent in SpellStore spell #%u request.", spell_id);
-        }
-
-        return false;
-    }
-
-    PetSpellMap::iterator itr = m_spells.find(spell_id);
-    if (itr != m_spells.end())
-    {
-        if (itr->second.state == PETSPELL_REMOVED)
-        {
-            m_spells.erase(itr);
-            state = PETSPELL_CHANGED;
-        }
-        else if (state == PETSPELL_UNCHANGED && itr->second.state != PETSPELL_UNCHANGED)
-        {
-            // can be in case spell loading but learned at some previous spell loading
-            itr->second.state = PETSPELL_UNCHANGED;
-
-            if (active == ACT_ENABLED)
-            {
-                ToggleAutocast(spell_id, true);
-            }
-            else if (active == ACT_DISABLED)
-            {
-                ToggleAutocast(spell_id, false);
-            }
-
-            return false;
-        }
-        else
-        {
-            return false;
-        }
-    }
-
-    uint32 oldspell_id = 0;
-
-    PetSpell newspell;
-    newspell.state = state;
-    newspell.type = type;
-
-    if (active == ACT_DECIDE)                               // active was not used before, so we save it's autocast/passive state here
-    {
-        if (IsPassiveSpell(spellInfo))
-        {
-            newspell.active = ACT_PASSIVE;
-        }
-        else
-        {
-            newspell.active = ACT_DISABLED;
-        }
-    }
-    else
-    {
-        newspell.active = active;
-    }
-
-    // talent: unlearn all other talent ranks (high and low)
-    if (TalentSpellPos const* talentPos = GetTalentSpellPos(spell_id))
-    {
-        if (TalentEntry const* talentInfo = sTalentStore.LookupEntry(talentPos->talent_id))
-        {
-            for (int i = 0; i < MAX_TALENT_RANK; ++i)
-            {
-                // skip learning spell and no rank spell case
-                uint32 rankSpellId = talentInfo->RankID[i];
-                if (!rankSpellId || rankSpellId == spell_id)
-                {
-                    continue;
-                }
-
-                // skip unknown ranks
-                if (!HasSpell(rankSpellId))
-                {
-                    continue;
-                }
-                removeSpell(rankSpellId, false, false);
-            }
-        }
-    }
-    else if (sSpellMgr.GetSpellRank(spell_id) != 0)
-    {
-        for (PetSpellMap::const_iterator itr2 = m_spells.begin(); itr2 != m_spells.end(); ++itr2)
-        {
-            if (itr2->second.state == PETSPELL_REMOVED)
-            {
-                continue;
-            }
-
-            if (sSpellMgr.IsRankSpellDueToSpell(spellInfo, itr2->first))
-            {
-                // replace by new high rank
-                if (sSpellMgr.IsHighRankOfSpell(spell_id, itr2->first))
-                {
-                    newspell.active = itr2->second.active;
-
-                    if (newspell.active == ACT_ENABLED)
-                    {
-                        ToggleAutocast(itr2->first, false);
-                    }
-
-                    oldspell_id = itr2->first;
-                    unlearnSpell(itr2->first, false, false);
-                    break;
-                }
-                // ignore new lesser rank
-                else if (sSpellMgr.IsHighRankOfSpell(itr2->first, spell_id))
-                {
-                    return false;
-                }
-            }
-        }
-    }
-
-    m_spells[spell_id] = newspell;
-
-    if (IsPassiveSpell(spellInfo))
-    {
-        CastSpell(this, spell_id, true);
-    }
-    else
-    {
-        m_charmInfo->AddSpellToActionBar(spell_id, ActiveStates(newspell.active));
-    }
-
-    if (newspell.active == ACT_ENABLED)
-    {
-        ToggleAutocast(spell_id, true);
-    }
-
-    uint32 talentCost = GetTalentSpellCost(spell_id);
-    if (talentCost)
-    {
-        m_usedTalentCount += talentCost;
-        UpdateFreeTalentPoints(false);
-    }
-    return true;
-}
-
-/**
- * @brief Learns a spell for the pet.
- *
- * @param spell_id The spell to learn.
- * @return true if the spell was learned; otherwise, false.
- */
-bool Pet::learnSpell(uint32 spell_id)
-{
-    // prevent duplicated entires in spell book
-    if (!addSpell(spell_id))
-    {
-        return false;
-    }
-
-    if (!m_loading)
-    {
-        Unit* owner = GetOwner();
-        if (owner && owner->GetTypeId() == TYPEID_PLAYER)
-        {
-            WorldPacket data(SMSG_PET_LEARNED_SPELL, 4);
-            data << uint32(spell_id);
-            ((Player*)owner)->GetSession()->SendPacket(&data);
-
-            {
-                ((Player*)owner)->PetSpellInitialize();
-            }
-        }
-    }
-    return true;
-}
-
-void Pet::InitLevelupSpellsForLevel()
-{
-    uint32 level = getLevel();
-
-    if (PetLevelupSpellSet const* levelupSpells = GetCreatureInfo()->Family ? sSpellMgr.GetPetLevelupSpellList(GetCreatureInfo()->Family) : NULL)
-    {
-        // PetLevelupSpellSet ordered by levels, process in reversed order
-        for (PetLevelupSpellSet::const_reverse_iterator itr = levelupSpells->rbegin(); itr != levelupSpells->rend(); ++itr)
-        {
-            // will called first if level down
-            if (itr->first > level)
-            {
-                unlearnSpell(itr->second, true);            // will learn prev rank if any
-            }
-            // will called if level up
-            else
-            {
-                learnSpell(itr->second);                    // will unlearn prev rank if any
-            }
-        }
-    }
-
-    int32 petSpellsId = GetCreatureInfo()->PetSpellDataId ? -(int32)GetCreatureInfo()->PetSpellDataId : GetEntry();
-
-    // default spells (can be not learned if pet level (as owner level decrease result for example) less first possible in normal game)
-    if (PetDefaultSpellsEntry const* defSpells = sSpellMgr.GetPetDefaultSpellsEntry(petSpellsId))
-    {
-        for (int i = 0; i < MAX_CREATURE_SPELL_DATA_SLOT; ++i)
-        {
-            SpellEntry const* spellEntry = sSpellStore.LookupEntry(defSpells->spellid[i]);
-            if (!spellEntry)
-            {
-                continue;
-            }
-
-            // will called first if level down
-            if (spellEntry->GetSpellLevel() > level)
-            {
-                unlearnSpell(spellEntry->Id, true);
-            }
-            // will called if level up
-            else
-            {
-                learnSpell(spellEntry->Id);
-            }
-        }
-    }
-}
-
-/**
- * @brief Unlearns a pet spell.
- *
- * @param spell_id The spell to remove.
- * @param learn_prev true to relearn the previous rank.
- * @param clear_ab true to clear the action bar slot when needed.
- * @return true if the spell was removed; otherwise, false.
- */
-bool Pet::unlearnSpell(uint32 spell_id, bool learn_prev, bool clear_ab)
-{
-    if (removeSpell(spell_id, learn_prev, clear_ab))
-    {
-        if (!m_loading)
-        {
-            if (Unit* owner = GetOwner())
-            {
-                if (owner->GetTypeId() == TYPEID_PLAYER)
-                {
-                    WorldPacket data(SMSG_PET_REMOVED_SPELL, 4);
-                    data << uint32(spell_id);
-                    ((Player*)owner)->GetSession()->SendPacket(&data);
-                }
-            }
-        }
-        return true;
-    }
-    return false;
-}
-
-/**
- * @brief Removes a spell from the pet spellbook.
- *
- * @param spell_id The spell to remove.
- * @param learn_prev true to relearn the previous rank.
- * @param clear_ab true to clear the action bar slot when needed.
- * @return true if the spell was removed; otherwise, false.
- */
-bool Pet::removeSpell(uint32 spell_id, bool learn_prev, bool clear_ab)
-{
-    PetSpellMap::iterator itr = m_spells.find(spell_id);
-    if (itr == m_spells.end())
-    {
-        return false;
-    }
-
-    if (itr->second.state == PETSPELL_REMOVED)
-    {
-        return false;
-    }
-
-    if (itr->second.state == PETSPELL_NEW)
-    {
-        m_spells.erase(itr);
-    }
-    else
-    {
-        itr->second.state = PETSPELL_REMOVED;
-    }
-
-    RemoveAurasDueToSpell(spell_id);
-
-    uint32 talentCost = GetTalentSpellCost(spell_id);
-    if (talentCost > 0)
-    {
-        if (m_usedTalentCount > talentCost)
-        {
-            m_usedTalentCount -= talentCost;
-        }
-        else
-        {
-            m_usedTalentCount = 0;
-        }
-
-        UpdateFreeTalentPoints(false);
-    }
-
-    if (learn_prev)
-    {
-        if (uint32 prev_id = sSpellMgr.GetPrevSpellInChain(spell_id))
-        {
-            learnSpell(prev_id);
-        }
-        else
-        {
-            learn_prev = false;
-        }
-    }
-
-    // if remove last rank or non-ranked then update action bar at server and client if need
-    if (clear_ab && !learn_prev && m_charmInfo->RemoveSpellFromActionBar(spell_id))
-    {
-        if (!m_loading)
-        {
-            // need update action bar for last removed rank
-            if (Unit* owner = GetOwner())
-                if (owner->GetTypeId() == TYPEID_PLAYER)
-                {
-                    ((Player*)owner)->PetSpellInitialize();
-                }
-        }
-    }
-
-    return true;
-}
-
-/**
- * @brief Removes unknown spells from the pet action bar.
- */
-void Pet::CleanupActionBar()
-{
-    for (int i = 0; i < MAX_UNIT_ACTION_BAR_INDEX; ++i)
-        if (UnitActionBarEntry const* ab = m_charmInfo->GetActionBarEntry(i))
-            if (uint32 action = ab->GetAction())
-                if (ab->IsActionBarForSpell() && !HasSpell(action))
-                {
-                    m_charmInfo->SetActionBar(i, 0, ACT_DISABLED);
-                }
-}
-
-/**
- * @brief Initializes the pet spellbook and action bar for a newly created pet.
- */
-void Pet::InitPetCreateSpells()
-{
-    m_charmInfo->InitPetActionBar();
-    m_spells.clear();
-
-    LearnPetPassives();
-
-    CastPetAuras(false);
-}
-
-bool Pet::resetTalents(bool no_cost)
-{
-    Unit* owner = GetOwner();
-    if (!owner || owner->GetTypeId() != TYPEID_PLAYER)
-    {
-        return false;
-    }
-
-    // not need after this call
-    if (((Player*)owner)->HasAtLoginFlag(AT_LOGIN_RESET_PET_TALENTS))
-    {
-        ((Player*)owner)->RemoveAtLoginFlag(AT_LOGIN_RESET_PET_TALENTS, true);
-    }
-
-    CreatureInfo const* ci = GetCreatureInfo();
-    if (!ci)
-    {
-        return false;
-    }
-    // Check pet talent type
-    CreatureFamilyEntry const* pet_family = sCreatureFamilyStore.LookupEntry(ci->Family);
-    if (!pet_family || pet_family->petTalentType < 0)
-    {
-        return false;
-    }
-
-    Player* player = (Player*)owner;
-
-    if (m_usedTalentCount == 0)
-    {
-        UpdateFreeTalentPoints(false);                      // for fix if need counter
-        return false;
-    }
-
-    uint32 cost = 0;
-
-    if (!no_cost)
-    {
-        cost = resetTalentsCost();
-
-        if (player->GetMoney() < cost)
-        {
-            player->SendBuyError(BUY_ERR_NOT_ENOUGHT_MONEY, 0, 0, 0);
-            return false;
-        }
-    }
-
-    for (unsigned int i = 0; i < sTalentStore.GetNumRows(); ++i)
-    {
-        TalentEntry const* talentInfo = sTalentStore.LookupEntry(i);
-
-        if (!talentInfo) continue;
-
-        TalentTabEntry const* talentTabInfo = sTalentTabStore.LookupEntry(talentInfo->TalentTab);
-
-        if (!talentTabInfo)
-        {
-            continue;
-        }
-
-        // unlearn only talents for pets family talent type
-        if (!((1 << pet_family->petTalentType) & talentTabInfo->petTalentMask))
-        {
-            continue;
-        }
-
-        for (int j = 0; j < MAX_TALENT_RANK; ++j)
-            if (talentInfo->RankID[j])
-            {
-                removeSpell(talentInfo->RankID[j], !IsPassiveSpell(talentInfo->RankID[j]), false);
-            }
-    }
-
-    UpdateFreeTalentPoints(false);
-
-    if (!no_cost)
-    {
-        player->ModifyMoney(-(int64)cost);
-
-        m_resetTalentsCost = cost;
-        m_resetTalentsTime = time(NULL);
-    }
-    player->PetSpellInitialize();
-    return true;
-}
-
-void Pet::resetTalentsForAllPetsOf(Player* owner, Pet* online_pet /*= NULL*/)
-{
-    // not need after this call
-    if (((Player*)owner)->HasAtLoginFlag(AT_LOGIN_RESET_PET_TALENTS))
-    {
-        ((Player*)owner)->RemoveAtLoginFlag(AT_LOGIN_RESET_PET_TALENTS, true);
-    }
-
-    // reset for online
-    if (online_pet)
-    {
-        online_pet->resetTalents(true);
-    }
-
-    // now need only reset for offline pets (all pets except online case)
-    uint32 except_petnumber = online_pet ? online_pet->GetCharmInfo()->GetPetNumber() : 0;
-
-    QueryResult* resultPets = CharacterDatabase.PQuery(
-                                  "SELECT `id` FROM `character_pet` WHERE `owner` = '%u' AND `id` <> '%u'",
-                                  owner->GetGUIDLow(), except_petnumber);
-
-    // no offline pets
-    if (!resultPets)
-    {
-        return;
-    }
-
-    QueryResult* result = CharacterDatabase.PQuery(
-                              "SELECT DISTINCT `pet_spell`.`spell` FROM `pet_spell`, `character_pet` "
-                              "WHERE `character_pet`.`owner` = '%u' AND `character_pet`.`id` = `pet_spell`.`guid` AND `character_pet`.`id` <> %u",
-                              owner->GetGUIDLow(), except_petnumber);
-
-    if (!result)
-    {
-        delete resultPets;
-        return;
-    }
-
-    bool need_comma = false;
-    std::ostringstream ss;
-    ss << "DELETE FROM `pet_spell` WHERE `guid` IN (";
-
-    do
-    {
-        Field* fields = resultPets->Fetch();
-
-        uint32 id = fields[0].GetUInt32();
-
-        if (need_comma)
-        {
-            ss << ",";
-        }
-
-        ss << id;
-
-        need_comma = true;
-    }
-    while (resultPets->NextRow());
-
-    delete resultPets;
-
-    ss << ") AND `spell` IN (";
-
-    bool need_execute = false;
-    do
-    {
-        Field* fields = result->Fetch();
-
-        uint32 spell = fields[0].GetUInt32();
-
-        if (!GetTalentSpellCost(spell))
-        {
-            continue;
-        }
-
-        if (need_execute)
-        {
-            ss << ",";
-        }
-
-        ss << spell;
-
-        need_execute = true;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    if (!need_execute)
-    {
-        return;
-    }
-
-    ss << ")";
-
-    CharacterDatabase.Execute(ss.str().c_str());
-}
-
-void Pet::UpdateFreeTalentPoints(bool resetIfNeed)
-{
-    uint32 level = getLevel();
-    uint32 talentPointsForLevel = GetMaxTalentPointsForLevel(level);
-    // Reset talents in case low level (on level down) or wrong points for level (hunter can unlearn TP increase talent)
-    if (talentPointsForLevel == 0 || m_usedTalentCount > talentPointsForLevel)
-    {
-        // Remove all talent points (except for admin pets)
-        if (resetIfNeed)
-        {
-            Unit* owner = GetOwner();
-            if (!owner || owner->GetTypeId() != TYPEID_PLAYER || ((Player*)owner)->GetSession()->GetSecurity() < SEC_ADMINISTRATOR)
-            {
-                resetTalents(true);
-            }
-            else
-            {
-                SetFreeTalentPoints(0);
-            }
-        }
-        else
-        {
-            SetFreeTalentPoints(0);
-        }
-    }
-    else
-    {
-        SetFreeTalentPoints(talentPointsForLevel - m_usedTalentCount);
-    }
-}
-
-void Pet::InitTalentForLevel()
-{
-    UpdateFreeTalentPoints();
-
-    Unit* owner = GetOwner();
-    if (!owner || owner->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (!m_loading)
-    {
-        ((Player*)owner)->SendTalentsInfoData(true);
-    }
-}
-
-/**
- * @brief Computes the current pet talent reset cost.
- *
- * @return The reset cost in copper.
- */
-uint32 Pet::resetTalentsCost() const
-{
-    uint32 days = uint32(sWorld.GetGameTime() - m_resetTalentsTime) / DAY;
-
-    // The first time reset costs 10 silver; after 1 day cost is reset to 10 silver
-    if (m_resetTalentsCost < 10 * SILVER || days > 0)
-    {
-        return 10 * SILVER;
-    }
-    // then 50 silver
-    else if (m_resetTalentsCost < 50 * SILVER)
-    {
-        return 50 * SILVER;
-    }
-    // then 1 gold
-    else if (m_resetTalentsCost < 1 * GOLD)
-    {
-        return 1 * GOLD;
-    }
-    // then increasing at a rate of 1 gold; cap 10 gold
-    else
-    {
-        return (m_resetTalentsCost + 1 * GOLD > 10 * GOLD ? 10 * GOLD : m_resetTalentsCost + 1 * GOLD);
-    }
-}
-
-uint8 Pet::GetMaxTalentPointsForLevel(uint32 level)
-{
-    uint8 points = (level >= 20) ? ((level - 16) / 4) : 0;
-    // Mod points from owner SPELL_AURA_MOD_PET_TALENT_POINTS
-    if (Unit* owner = GetOwner())
-    {
-        points += owner->GetTotalAuraModifier(SPELL_AURA_MOD_PET_TALENT_POINTS);
-    }
-    return points;
-}
-
-/**
- * @brief Enables or disables autocast for a pet spell.
- *
- * @param spellid The spell to update.
- * @param apply true to enable autocast; false to disable it.
- */
-void Pet::ToggleAutocast(uint32 spellid, bool apply)
-{
-    if (IsPassiveSpell(spellid))
-    {
-        return;
-    }
-
-    PetSpellMap::iterator itr = m_spells.find(spellid);
-    PetSpell &petSpell = itr->second;
-
-    uint32 i;
-
-    if (apply)
-    {
-        for (i = 0; i < m_autospells.size() && m_autospells[i] != spellid; ++i)
-        {
-            ;                                                // just search
-        }
-
-        if (i == m_autospells.size())
-        {
-            m_autospells.push_back(spellid);
-
-            if (petSpell.active != ACT_ENABLED)
-            {
-                petSpell.active = ACT_ENABLED;
-                if (petSpell.state != PETSPELL_NEW)
-                {
-                    petSpell.state = PETSPELL_CHANGED;
-                }
-            }
-        }
-    }
-    else
-    {
-        AutoSpellList::iterator itr2 = m_autospells.begin();
-        for (i = 0; i < m_autospells.size() && m_autospells[i] != spellid; ++i, ++itr2)
-        {
-            ;                                                // just search
-        }
-
-        if (i < m_autospells.size())
-        {
-            m_autospells.erase(itr2);
-            if (petSpell.active != ACT_DISABLED)
-            {
-                petSpell.active = ACT_DISABLED;
-                if (petSpell.state != PETSPELL_NEW)
-                {
-                    petSpell.state = PETSPELL_CHANGED;
-                }
-            }
-        }
-    }
-}
 
 /**
  * @brief Creates the pet world object from creature data.
